requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0